upcoming: [DI-28682] - Fix UT
