+++ packages/manager/src/features/CloudPulse/Widget/components/CloudPulseLineGraph.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { AreaChart } from 'src/components/AreaChart/AreaChart';
 import { useFlags } from 'src/hooks/useFlags';
 
-import { humanizeLargeData } from '../../Utils/CloudPulseWidgetUtils';
+import { humanizeLargeData } from '../../Utils/utils';
 
 import type { AreaChartProps } from 'src/components/AreaChart/AreaChart';
 
@@ -33,9 +33,11 @@ export const CloudPulseLineGraph = React.memo((props: CloudPulseLineGraph) => {
 
   const noDataMessage = 'No data to display';
   const isHumanizableUnit =
-    flags.aclp?.humanizableUnits?.some(
+    flags.aclp?.humanizableUnits ??
+    ['Count']?.some(
       (unitElement) => unitElement.toLowerCase() === unit.toLowerCase()
-    ) ?? false;
+    ) ??
+    false;
   return (
     <Box
       sx={{
